Set title in screen even when using Apple Terminal
In modules/terminal/init.zsh, the auto-title code currently will only
set the screen title if $TERM_PROGRAM != 'Apple_Terminal'. As a result,
if you use screen within an Apple Terminal, the screen title will not be
correctly set.

To fix this, I changed the logic such that the screen title will be set
even if the $TERM_PROGRAM == 'Apple Terminal'.